requirements.txt
albumentations==1.3.0
h5py==3.1.0
matplotlib==3.6.2
numpy==1.22.3
pandas==1.5.2
Pillow==9.2.0
pytorch_lightning==1.8.6
pytorch_ssim==0.1
PyYAML==6.0
Requests==2.30.0
setuptools==65.5.0
thop==0.1.1.post2209072238
timm==0.6.12
torch==1.13.0
torchsummary==1.5.1
torchvision==0.14.0
